fix: product import upsert + price field decimal input
- ProductImport: rewrite to ToCollection with 2-pass validation then upsert
  keyed on sku column (no more duplicate SP-xxxx on re-import)
- ProductController: show created/updated count on import success
- OrderController, QuotationController: discount_amount numeric (not integer)
- ProjectWipEntry: cast amount as decimal:2 (not integer)
- All price/unit_price fields: add step="any" so browser accepts decimals
  (Orders, Quotations, PurchaseOrders, StockEntries, StockExits,
   PurchaseReturns, Services, PriceLists, Employees salary fields)

+++ app/Http/Controllers/Sales/OrderController.php
@@ -105,7 +105,7 @@ class OrderController extends Controller
             'items.*.unit_price'          => ['required', 'numeric', 'min:0'],
             'items.*.vat_rate'            => ['nullable', 'numeric', 'min:0', 'max:100'],
             'items.*.discount_percent'    => ['nullable', 'numeric', 'min:0', 'max:100'],
-            'items.*.discount_amount'     => ['nullable', 'integer', 'min:0'],
+            'items.*.discount_amount'     => ['nullable', 'numeric', 'min:0'],
         ]);
 
         $isFdi = Customer::where('id', $data['customer_id'])->value('is_fdi');
@@ -293,7 +293,7 @@ class OrderController extends Controller
             'items.*.unit_price'       => ['required', 'numeric', 'min:0'],
             'items.*.vat_rate'         => ['nullable', 'numeric', 'min:0', 'max:100'],
             'items.*.discount_percent' => ['nullable', 'numeric', 'min:0', 'max:100'],
-            'items.*.discount_amount'  => ['nullable', 'integer', 'min:0'],
+            'items.*.discount_amount'  => ['nullable', 'numeric', 'min:0'],
         ]);
 
         $order->update([
